#71 💷 refactor: 커스텀 hooks useTrading 생성
